+++ klibs/KLEventQueue.py
@@ -22,10 +22,10 @@ from sdl2 import SDL_PumpEvents, SDL_FlushEvents, SDL_FIRSTEVENT, SDL_LASTEVENT
 from sdl2.ext import get_events
 
 
-def pump(return_events=False):
-    """Updates (and optionally retrieves) the contents of the input event queue.
+def pump(return_events=True):
+    """Retrieves the current contents of the input event queue.
     
-    In general, ``pump(True)`` is usually called at the start of a loop checking
+    In general, ``pump()`` is usually called at the start of a loop checking
     for input::
 
         flush() # Clear the input queue before starting the loop
@@ -33,33 +33,25 @@ def pump(return_events=False):
         start_time = precise_time()
         while (precise_time() - start_time) < 5.0 and not response:
 
-            q = pump(True) # Retrieve the current input queue contents
+            q = pump() # Retrieve the current input queue contents
 
             if key_pressed('z', queue=q):
                 response = 'left'
             elif key_pressed('/', queue=q):
                 response = 'right'
 
-    Note that calling ``pump(True)`` both returns and clears the contents of the
+    Note that calling ``pump()`` both returns and clears the contents of the
     event queue, so make sure to call it only once per loop to avoid missing
     input events.
 
-    Calling ``pump(False)`` will update the current state of different input
-    devices (e.g. the mouse cursor position, joystick button/axis states) but
-    will not fetch or clear any events from the queue.
-
     Args:
-        return_events (bool): If True, returns the contents of the input event queue.
+        return_events (bool): Deprecated, should always be True.
 
     Returns:
-        list or None: A list of SDL_Event objects if return_events is True,
-        otherwise ``None``.
+        list: A list of ``SDL_Event`` objects.
 
     """
-    # NOTE: get_events() empties queue, SDL_PumpEvents() doesn't
-    SDL_PumpEvents()
-    if return_events:
-        return get_events()
+    return get_events()
 
 
 def flush():